debug: use OrderedDict when dump to preserve the order
In Python 3.6+, dict order is kept in general, but before 3.6 it is
not preserved, and we must use OrderedDict to preserve the order when
MarshalDump.

YAML output is written yet as a dict for simplifity, but in future
we may change it to omap-format to preserve the order in the parser.